requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-groq
langchain-community
newsapi-python
newspaper3k
lxml_html_clean
pydantic
tavily-python
python-dotenv
httpx
requests
